HelloWorld/UVHelloWorld/pyproject.toml
[project]
name = "uvhelloworld"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ahmerarman", email = "ahmer.arman40@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "nest-asyncio>=1.6.0",
    "openai>=1.66.3",
    "openai-agents>=0.0.3",
    "python-dotenv>=1.0.1",
]

[project.scripts]
uvhelloworld = "uvhelloworld:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
